requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.3
numpy==1.24.2
pandas==1.5.3
pathlib==1.0.1
protobuf==4.22.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
